fixed a compiler warning about an implicit integer type conversion
Former-commit-id: 08b0be0138ec4f2e7b3a5c62e6e4b503a8dfff97
Former-commit-id: 95274b42eb68dc5cc96c8849c9e72150bffc1197

+++ SimCode/dynamics/_GeneralModuleFiles/stateData.h
@@ -42,8 +42,8 @@ public:
     Eigen::MatrixXd getState() const {return state;}
     Eigen::MatrixXd getStateDeriv() const {return stateDeriv;}
     std::string getName() const {return stateName;}
-    uint32_t getRowSize() const {return(state.innerSize());}
-    uint32_t getColumnSize() const {return(state.outerSize());}
+    uint32_t getRowSize() const {return((uint32_t)state.innerSize());}
+    uint32_t getColumnSize() const {return((uint32_t)state.outerSize());}
     bool isStateActive() {return stateEnabled;}
     void disable() {stateEnabled = false;}
     void enable() {stateEnabled = true;}
